refactor: Align package structure with core worker module hierarchy
