+++ core/worker/api/src/main/java/com/metasearch/android/core/worker/api/usecase/WorkerStatusUseCase.kt
@@ -1,4 +1,4 @@
-package com.metasearch.android.api.usecase
+package com.metasearch.android.core.worker.api.usecase
 
 import androidx.work.WorkInfo
 import kotlinx.coroutines.flow.Flow
