+++ feature/home/src/main/java/com/metasearch/android/feature/home/HomePresenter.kt
@@ -11,8 +11,8 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.geometry.Offset
 import androidx.paging.cachedIn
 import androidx.work.WorkInfo
-import com.metasearch.android.api.usecase.WorkScheduleUseCase
-import com.metasearch.android.api.usecase.WorkerStatusUseCase
+import com.metasearch.android.core.worker.api.usecase.WorkScheduleUseCase
+import com.metasearch.android.core.worker.api.usecase.WorkerStatusUseCase
 import com.metasearch.android.core.data.api.repository.GalleryRepository
 import com.metasearch.android.core.data.api.repository.PersonRepository
 import com.metasearch.android.core.model.Person
